requirements.txt
# Python version: 3.9.7
Flask==2.1.2
Flask-SocketIO==5.2.0
eventlet==0.33.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
SQLAlchemy==1.4.36
alembic==1.7.5
pytest==6.2.5
pytest-flask==1.2.0
flask-cors==3.0.10

